requirements.txt
certifi==2022.6.15
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.3
Flask_Cors==3.0.10
Flask_Login==0.6.2
Flask_WTF==1.0.1
mongoengine==0.24.2
playsound==1.3.0
python-dotenv==0.20.0
Werkzeug==2.2.0
WTForms==3.0.1
pillow==9.2.0
qrcode==7.3.1
mongomock==4.1.2
WebTest==3.0.0
coverage==6.4.4
webdriver-manager==3.8.3
selenium==4.4.3
boto3==1.29.6